# Conflicts:
#	logs/application.log

+++ src/main/java/com/etz/fraudeagleeyemanager/service/RuleService.java
@@ -6,9 +6,7 @@ import com.etz.fraudeagleeyemanager.dto.request.MapRuleToProductRequest;
 import com.etz.fraudeagleeyemanager.dto.request.UpdateMapRuleToProductRequest;
 import com.etz.fraudeagleeyemanager.dto.request.UpdateRuleRequest;
 import com.etz.fraudeagleeyemanager.dto.response.*;
-import com.etz.fraudeagleeyemanager.entity.ProductEntity;
-import com.etz.fraudeagleeyemanager.entity.ProductRule;
-import com.etz.fraudeagleeyemanager.entity.Rule;
+import com.etz.fraudeagleeyemanager.entity.*;
 import com.etz.fraudeagleeyemanager.exception.FraudEngineException;
 import com.etz.fraudeagleeyemanager.exception.ResourceNotFoundException;
 import com.etz.fraudeagleeyemanager.redisrepository.ProductRuleRedisRepository;
@@ -37,7 +35,7 @@ import java.util.*;
 public class RuleService {
 
 	@Autowired @Qualifier("redisTemplate")
-	private RedisTemplate<String, Object> fraudEngineRedisTemplate;
+	private final RedisTemplate<String, Object> fraudEngineRedisTemplate;
 	
 	@Autowired
 	RuleRedisRepository ruleRedisRepository;
@@ -49,7 +47,7 @@ public class RuleService {
 	RuleRepository ruleRepository;
 
 	@PersistenceContext
-	private EntityManager em;
+	private final EntityManager em;
 
 
 	@Autowired
@@ -61,7 +59,12 @@ public class RuleService {
 	@Autowired
 	EmailGroupRepository emailGroupRepository;
 
-	
+	public RuleService(RedisTemplate<String, Object> fraudEngineRedisTemplate, EntityManager em) {
+		this.fraudEngineRedisTemplate = fraudEngineRedisTemplate;
+		this.em = em;
+	}
+
+
 	public Rule createRule(CreateRuleRequest request) {
 		Rule ruleEntity = new Rule();
 		ruleEntity.setName(request.getRuleName());
@@ -140,8 +143,8 @@ public class RuleService {
 	}
 
 	public boolean deleteRule(Long parameterId) {
-		Optional<ProductRule> prodRuleEntity = productRuleRepository.findByRuleId(parameterId);
-		if (prodRuleEntity.isPresent()){
+		List<ProductRule> prodRuleEntity = productRuleRepository.findByRuleId(parameterId);
+		if (!(prodRuleEntity.isEmpty())){
 			throw new FraudEngineException("Please unmap the ruleId before deleting " + parameterId);
 		}
 		ruleRepository.deleteById(parameterId);
@@ -202,38 +205,52 @@ public class RuleService {
 		return createdProductRuleEntity;
 	}
 
-	public ProductRuleResponse updateProductRule(UpdateMapRuleToProductRequest request) {
-		ProductRule prodRuleEntity = productRuleRepository.findById(request.getProductRuleId())
-				.orElseThrow(() -> new ResourceNotFoundException("ProductRule Not found for Id " + request.getProductRuleId() ));
-
+	public List<ProductRuleResponse> updateProductRule(UpdateMapRuleToProductRequest request) {
+		List<ProductRule> prodRuleEntityList = productRuleRepository.findByRuleId(request.getProductRuleId());
+		if(prodRuleEntityList.isEmpty()){
+			throw new FraudEngineException("ProductRule Not found for Id " + request.getProductRuleId());
+		}
+		List<ProductRule> updatedProductRuleEntity = new ArrayList<>();
 		if (request.getEmailGroupId() != null){
 			emailGroupRepository.findById(request.getEmailGroupId()).orElseThrow(() ->
 					new ResourceNotFoundException("Email Group Id Not found for  Id " + request.getEmailGroupId() ));
 		}
-		prodRuleEntity.setNotifyAdmin(request.getNotifyAdmin());
-		prodRuleEntity.setEmailGroupId(request.getEmailGroupId());
-		prodRuleEntity.setNotifyCustomer(request.getNotifyCustomer());
-		prodRuleEntity.setStatus(request.getStatus());
-		prodRuleEntity.setAuthorised(request.getAuthorised());
-		prodRuleEntity.setUpdatedBy(request.getUpdatedBy());
-		ProductRule updatedEntity = productRuleRepository.save(prodRuleEntity);
+
+		for (ProductRule prodRuleEntity:prodRuleEntityList) {
+			prodRuleEntity.setNotifyAdmin(request.getNotifyAdmin());
+			prodRuleEntity.setEmailGroupId(request.getEmailGroupId());
+			prodRuleEntity.setNotifyCustomer(request.getNotifyCustomer());
+			prodRuleEntity.setStatus(request.getStatus());
+			prodRuleEntity.setAuthorised(request.getAuthorised());
+			prodRuleEntity.setUpdatedBy(request.getUpdatedBy());
+			ProductRule updatedEntity = productRuleRepository.save(prodRuleEntity);
+			updatedProductRuleEntity.add(updatedEntity);
+		}
 
 		//update Redis
 		productRuleRedisRepository.setHashOperations(fraudEngineRedisTemplate);
-		productRuleRedisRepository.update(prodRuleEntity);
-		return outputProductRuleResponseList(updatedEntity);
+		for (ProductRule prodRuleEntity:updatedProductRuleEntity) {
+			productRuleRedisRepository.update(prodRuleEntity);
+		}
+		return outputProductRuleResponseList(updatedProductRuleEntity);
 	}
 
-	public boolean deleteProductRule(Long productRuleId) {
-		ProductRule prodRuleEntity = productRuleRepository.findByRuleId(productRuleId)
-				.orElseThrow(() -> new ResourceNotFoundException("ProductRule Not found for productRuleId " + productRuleId ));
-		String redisId = prodRuleEntity.getProductCode()+ ":"+prodRuleEntity.getRuleId();
-		productRuleRepository.deleteByRuleId(productRuleId);
+	public boolean deleteProductRule(Long ruleId, String code) {
+		if (ruleId == null && code == null){
+			throw new FraudEngineException("Please enter value for code and ruleId");
+		}
+		Optional<ProductRule> productRuleEntity = productRuleRepository.findById(new ProductRuleId(ruleId,code));
 		productRuleRedisRepository.setHashOperations(fraudEngineRedisTemplate);
-		productRuleRedisRepository.delete(redisId);
+
+		if(productRuleEntity.isPresent()){
+			String redisId = code + ":"+ruleId;
+			productRuleRepository.deleteByRuleId(ruleId, code);
+			productRuleRedisRepository.delete(redisId);
+		}else {
+			throw new ResourceNotFoundException("ProductRule Not found for ruleId " + ruleId + " and code " + code);
+		}
 		return true;
 	}
-
 
 
 	public List<RuleProductResponse> getRuleProduct(String code) {
@@ -263,11 +280,14 @@ public class RuleService {
 		return ruleResponseList;
 	}
 
-	private ProductRuleResponse outputProductRuleResponseList(ProductRule productRule){
-
+	private List<ProductRuleResponse> outputProductRuleResponseList(List<ProductRule> productRuleList){
+		List<ProductRuleResponse> productRules = new ArrayList<>();
+		productRuleList.forEach( productRule -> {
 			ProductRuleResponse ruleResponse = new ProductRuleResponse();
 			BeanUtils.copyProperties(productRule,ruleResponse,"productEntity", "emailGroup", "rule");
-			return ruleResponse;
+			productRules.add(ruleResponse);
+		});
+		return productRules;
 	}
